Reorganized index router
