+++ src/main/index.ts
@@ -1,57 +1,24 @@
 import * as bodyParser from 'body-parser';
 import * as cookieParser from 'cookie-parser';
 import * as express from 'express';
-import { ObjectId } from 'mongodb';
 import * as path from 'path';
 import DatabaseConnector from './dao/DatabaseConnector';
-import UserDAO from './dao/UserDAO';
+import apiRouter from './routes/apiRouter';
 
 const app = express();
 const PORT = process.env.PORT || 3000;
 
 DatabaseConnector.open();
 
+app.set('view engine', 'pug');
+
 app.use(express.static(path.resolve('./static/public')));
 app.use(bodyParser({ extended: true }));
 app.use(cookieParser());
-
-app.set('view engine', 'pug');
+app.use('/api', apiRouter);
 
 app.get('/', (_, res) => {
     res.render('index');
-});
-
-app.post('/api/user/create', (req, res) => {
-    const dao = new UserDAO();
-    dao.createUser(req.body)
-        .then(() =>
-            res.send({
-                success: true,
-            })
-        )
-        .catch(err => res.send(new ErrorAPIResponse(err)));
-});
-
-app.post('/api/user/login', (req, res) => {
-    const dao = new UserDAO();
-    dao.authenticateLogin(req.body)
-        .then(auth => {
-            res.cookie('uid', auth.uid);
-            res.cookie('auth', auth.key);
-            res.send({ success: true, auth });
-        })
-        .catch(err => res.send(new ErrorAPIResponse(err)));
-});
-
-app.get('/api/user/authenticate', (req, res) => {
-    const dao = new UserDAO();
-    dao.authenticateKey(new ObjectId(req.cookies.uid), req.cookies.auth)
-        .then(passed => {
-            console.log(passed);
-            if (passed) res.send({ success: true });
-            else res.send(new ErrorAPIResponse(new Error('Invalid certificate')));
-        })
-        .catch(error => res.send(new ErrorAPIResponse(error)));
 });
 
 app.listen(PORT, () => {
@@ -63,24 +30,3 @@ app.listen(PORT, () => {
       Website                http://localhost:${PORT}
     ▢----------------------▢--------------------------▢`);
 });
-
-interface APIResponse {
-    success: boolean;
-    error?: Error;
-}
-
-abstract class BaseAPIResponse implements APIResponse {
-    success: boolean;
-    error?: Error;
-
-    constructor(success: boolean, error?: Error) {
-        this.success = success;
-        this.error = error;
-    }
-}
-
-export class ErrorAPIResponse extends BaseAPIResponse {
-    constructor(error: Error | string) {
-        super(false, typeof error === 'string' ? new Error(error) : error);
-    }
-}
